aged-care-robot-demo/requirements.txt
# LiveKit + OpenAI Dependencies
livekit==1.0.25
livekit-api==1.1.0

# OpenAI
openai>=1.40.0

# Audio processing
pyroomacoustics==0.7.3
scipy==1.12.0
numpy>=1.26.0
soundfile==0.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment
python-dotenv==1.0.0

# Utilities
aiofiles==25.1.0
aiohttp==3.13.3
